pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "deepvibe-ai-hedge-fund"
version = "0.1.0"
description = "Standalone MAD / MRAT cross-sectional strategy (Alpaca OHLCV, SQLite, Dash)"
requires-python = ">=3.10"
dependencies = [
  "alpaca-py>=0.42.0",
  "dash>=2.14.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "plotly>=5.18.0",
  "python-dotenv>=1.0.0",
  "pytz>=2023.3",
  "requests>=2.28.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.1.0"]

[tool.setuptools.packages.find]
where = ["src"]
